requirements.txt
SmartApi-python
pandas
pandas_ta
python-dotenv
logzero
schedule
